requirements.txt
python-telegram-bot==21.0.1
yt-dlp==2024.3.10